requirements.txt
# Database
SQLAlchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Telegram Bot API (aiogram)
aiogram>=3.3.0

# Proxy support
aiohttp-socks>=0.8.4



# Environment Variables
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0



# Logging
loguru>=0.7.0